Global Ignore
May there be a command that works similar to /ignore would be applicable throughout the whole server network. Unfortunately, /ignore only works if both parties (the ignorer and the ignoree) are on the same server. However, as soon as one party joins a separate server on the BC network, the ignore no longer has any impact and no longer serves its function. Please create something to a similar effect. It would be very much appreciated.